% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreePlot.R
\name{ScreePlot}
\alias{ScreePlot}
\title{ScreePlot}
\usage{
ScreePlot(data, scaling = "NO")
}
\arguments{
\item{data}{A data frame containing the data matrix. The data frame  has to
have the same structure of the input data frame of Biplot3D and Biplot 2D
functions. See Biplot3D and Biplot 2D help pages for further details.}

\item{scaling}{Do you wish to scale your data prior to PC calculation?
Set "YES" if you do; otherwise set "NO". The latter choice is set as default.}
}
\value{
An interactive screeplot bar chart as a Plotly object.
}
\description{
Generates an interactive two-dimensional bar chart. The height
of each bar shows the variance explained by the corresponding principal
component.
}
\examples{
\dontrun{

> ScreePlot(data,scaling="NO")
# Here, a screeplot bar chart is generated after performing the Principal
Component Analysis (PCA) on centered data.

}

}
